Create SplashViewModel
