+++ app/build.gradle.kts
@@ -80,12 +80,14 @@ dependencies {
     implementation(libs.kotlinx.serialization.json)
     implementation(libs.lifecycle.runtime.ktx)
     implementation(libs.timber)
+    implementation(libs.hilt.navigation.compose)
 
     // Firebase
     implementation(platform(libs.firebase.bom))
     implementation(libs.firebase.analytics)
     implementation(libs.firebase.crashlytics)
     implementation(libs.firebase.perf)
+    implementation(project(":feature:trip-planner:state"))
 
     // Test
     androidTestImplementation(libs.test.androidxTestExtJunit)
